refactored messaging conversation information into its own view file
