+++ models/message.py
@@ -5,6 +5,7 @@ class Message(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     link = db.Column(db.String(500))
     image = db.Column(db.String(500))
+    sender = db.Column(db.String(500), nullable=False)
 
     def __init__(self, link, image):
         self.link = link
