+++ templates/conversation.html
@@ -57,6 +57,7 @@
         <div class="ui buttons">
             <form class="ui form" method="POST" enctype="multipart/form-data">
                 <input type=file name = "file" value="Upload">
+                <button class="ui button">Upload</button>
             </form>
             <button class="ui primary button">Select</button>
             <button class="ui button">Random</button>
